requirements.txt
asgiref==3.4.1
Django==3.2.11
django-environ==0.8.1
Pillow==9.0.0
psycopg2==2.9.3
python-dotenv==0.19.2
pytz==2021.3
sqlparse==0.4.2
typing_extensions==4.0.1
